Fix coming in hot

+++ models/mysql/index.es6
@@ -8,11 +8,14 @@ const basename = path.basename(module.filename);
 const db = {};
 const sequelize = new Sequelize(mysqlConfig.database, mysqlConfig.username, mysqlConfig.password, mysqlConfig);
 
-/* This is for whenver we update schema, need to create a better solution */
-sequelize.query('SET FOREIGN_KEY_CHECKS=0')
-  .then(() => sequelize.sync({force: true}))
-  .then(() => sequelize.query('SET FOREIGN_KEY_CHECKS=1'));
-
+/* This is for whenver we update schema, need to create a better solution
+* We need to only run when not staging (aka not tests are running) since the tests
+* individually sync the database and if you do it here there are problems */
+if (process.env.NODE_ENV !== 'staging') {
+  sequelize.query('SET FOREIGN_KEY_CHECKS=0')
+    .then(() => sequelize.sync({force: true}))
+    .then(() => sequelize.query('SET FOREIGN_KEY_CHECKS=1'));
+}
 // TODO logging => console.tag('sequelize').bind(console)
 
 fs.readdirSync(__dirname)
